lam/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lam-refractor"
version = "0.1.0"
description = "Modular refactoring of LAM (Large Avatar Model)"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "Apache-2.0"}
authors = [
    {name = "LAM Refactor Contributors"}
]

keywords = ["avatar", "3d", "flame", "gaussian-splatting", "deep-learning"]

classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.20.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

full = [
    "pytorch3d",
    "trimesh",
    "PIL",
    "opencv-python",
]

[project.urls]
Homepage = "https://github.com/aigc3d/LAM"
Documentation = "https://github.com/aigc3d/LAM"
Repository = "https://github.com/aigc3d/LAM"

[tool.setuptools]
packages = ["lam"]

[tool.setuptools.package-data]
lam = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
